+++ src/Device/DeviceSelector/DeviceSelector.jsx
@@ -123,7 +123,6 @@ DeviceSelector.propTypes = {
         serialport: bool,
         jlink: bool,
     }).isRequired,
-    /* eslint-disable react/require-default-props */
     deviceSetup: exact({
         jprog: object,
         dfu: object,